requirements.txt
-f /usr/share/pip-wheels
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
django-crispy-forms==1.9.2
djangorestframework==3.12.2
pytz==2020.1
sqlparse==0.4.1
